requirements.txt
aiogram
openpyxl
pandas
python-dotenv
